+++ app/_shared/user/services/userAuthService/index.ts
@@ -1,0 +1,1 @@
+export * as userAuthService from "./service";
